requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
openai
langchain
langchain[community]
langchain-openai
faiss-cpu 
unstructured 
markdown